chore(code): Use Macros for HasFactory
This PR moves to using macros for the `fake` and `factory` methods found previously on the `HasFactory` trait. This means that those methods can safely be used in tests without using the trait, allowing the package to be omitted in production without PHP failing.
This PR also introduces docker images for quickly getting started contributing to the request factories package, and being able to run our test suite and other checks.